Day-2 (25/05/2025): Kernel Development
Information:
we have created 2 new files one with only init function and one with
only exit function, so there are 2 cases.
case-1: file only has init function
In this case this module is a non-removable module, and is a common
thing in linux kernel.
case-2: file only has exit function
In this case this module when loaded in kernel won't do any
initialization, but during the exit of module from kernel the exit
function for cleanup will be invoked and executed.

Steps:
1. Makefile was changed, as here we are compiling 3 file in a single
   Makefile, := this is used for initialization and += is used to add
   additional files.
2. we have merged the 2 .c files init and exit to make a kernel module
   that uses init of one file and exit of one file just by making the
   necessary changes in the Makefile. In the Makefile we have
   reinitialized using := cmd and given a file name that doesn't
   exist in the codebase. After that we initialized the file in the
   Makefile.
   cmd: obj-m := kernel_module_init_exit.ko (non-existant file)
   kernel_module_init_exit-objs := init.ko exit.ko (this tells that to
   create kernel_module_init_exit file we need a combination of these
   files).

tail -f /var/log/syslog: this shows the last 10 lines of the file, as 10
is the default value. the -f flag means follow, so as soon as a new
entry is made in this log file it will also be shown in the terminal.

MODULE_LICENSE("GPL"): If this line in missing kernel device drivers
then kernel treats it as if someone is trying to hack the kernel, or
break its functionality. Also if this line is missing certain kernel
functionality will not be given.

__init & __initdata: __init & __initdata is used for removing
the function and variable respectively from kernel RAM  after the module
initialization in the kernel. Which thereby reduces the kernel size.

Reviewed-By: Shivam Kumar