Day-6 (31/05/2025): Kernel Development
module_param: This macro helps us pass CLI parameters to the kernel
module, here I am passing it to a kernel module which is out-of-tree.
To pass CLI arguments to kernel modules, look below.
eg. sudo insmod params_module.ko count=5
here count the global variable which is passed to module_param macro,
important thing to note here is that we should always pass global
variables only to this macro. This macro also required import of an
additional header which is <linx/moduleparam.h>

Refer:
-> https://tldp.org/LDP/lkmpg/2.6/html/x323.html

Reviewed-By: Shivam Kumar