Day-7 (1/06/2025): Kernel Development
The Intel 80386 architecture has 4 of modes, which are called rings.
Unix uses only two rings; the highest ring (ring 0, also known as
`supervisor mode' where everything is allowed to happen) and the lowest
ring, which is called `user mode'.

The kernel has its own space of memory as well. Since a module is code
which can be dynamically inserted and removed in the kernel (as opposed
to a semi-autonomous object), it shares the kernel's codespace rather than
having its own. Therefore, if your module segfaults, the kernel segfaults.
And if you start writing over data because of an off-by-one error, then
you're trampling on kernel data (or code). I would like to point out that
the above discussion is true for any operating system which uses a monolithic
kernel. There are things called microkernels which have modules which get
their own codespace. The GNU Hurd and QNX Neutrino are two examples of a
microkernel.

/proc/kallsyms: contains symbols that have been exported by kernel.
The file /proc/kallsyms holds all the symbols that the kernel knows
about and which are therefore accessible to your modules since they
share the kernel's codespace.

Doc: https://tldp.org/LDP/lkmpg/2.6/html/x427.html#FTN.AEN467
Reviewed-By: