Day-9 (06/06/2025): Character Device Driver
Today I created a character device driver. We cannot randomly allocate a
major number, hence during register_chrdev() we give 0, indicating that we
want kernel to assign a dynamic major number to the device driver, the function
returns the major number, after that we should create device file, which could be
automated using udev, but I am using mknod manually to create that device file
as of now.

try_module_get (THIS_MODULE): this updated a counter value which is keeping
track of the processes that are using this module, whenever we do sudo rmmod
and it fails it’s because of this macro.

module_put (THIS_MODULE): this decreases the count of the counter, if the
counter value is 0, that means no process is using this module, hence sudo
rmmod is safe to perform.

Reviewed-By: Shivam Kumar