Day-10 (10/06/2025): Linear Buffer Character Driver
A Linear Buffer character driver is created, just like we have black
hole in the universe, we have /dev/null in linux, I have tried replicating
that character driver, which is named as /dev/sk_linear_buffer, which is
just a buffer, which can be used to store and read 1024 bytes of data.
/dev/null character device driver will have an almost similar coding, it
just that the device_read and device_write function might be a bit
different.

Reviewed-By: Shivam Kumar